requirements.txt
pandas
yfinance
scikit-learn
lightgbm
joblib
streamlit
plotly-express